added factory for users and updated posts factory where only authors with hairdresser role can post in the blog

+++ database/seeders/PostSeeder.php
@@ -3,9 +3,11 @@
 namespace Database\Seeders;
 
 use App\Models\Tag;
+use App\Models\User;
 use App\Models\Blog\Post;
 use App\Models\Blog\Comment;
 use Illuminate\Database\Seeder;
+use Illuminate\Support\Facades\File;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Database\Console\Seeds\WithoutModelEvents;
 
@@ -18,8 +20,10 @@ class PostSeeder extends Seeder
      */
     public function run()
     {
-        $path = public_path('uploads').'/blog';
-        delete_folder($path);
+        if (Storage::exists('users')) {
+            $path = public_path('uploads').'/blog';
+            delete_folder($path);
+        }
 
         $tags_id = Tag::pluck('id');
         $posts = Post::factory(20)->create();
@@ -29,7 +33,6 @@ class PostSeeder extends Seeder
 
             Comment::factory(3)->create([
                 'post_id' => $post->id,
-                'author_id' => rand(1, 4),
             ]);
         });
     }
